requirements.txt
pyfrc ~= 2020.1.4
robotpy-ctre ~= 2020.3.1
robotpy-rev ~= 2020.1.1
robotpy-rev-color ~= 2020.1.2.0
robotpy-wpilib-utilities ~= 2020.2.0
wpilib ~= 2020.3.2.4
pip
setuptools ~= 46.0.0